--unsafe-fixes

+++ youtube_dl/aes.py
@@ -32,7 +32,7 @@ def aes_ctr_decrypt(data, key, counter):
     @returns {int[]}           decrypted data
     """
     expanded_key = key_expansion(key)
-    block_count = int(ceil(float(len(data)) / BLOCK_SIZE_BYTES))
+    block_count = ceil(float(len(data)) / BLOCK_SIZE_BYTES)
 
     decrypted_data = []
     for i in range(block_count):
@@ -57,7 +57,7 @@ def aes_cbc_decrypt(data, key, iv):
     @returns {int[]}           decrypted data
     """
     expanded_key = key_expansion(key)
-    block_count = int(ceil(float(len(data)) / BLOCK_SIZE_BYTES))
+    block_count = ceil(float(len(data)) / BLOCK_SIZE_BYTES)
 
     decrypted_data = []
     previous_cipher_block = iv
@@ -83,7 +83,7 @@ def aes_cbc_encrypt(data, key, iv):
     @returns {int[]}           encrypted data
     """
     expanded_key = key_expansion(key)
-    block_count = int(ceil(float(len(data)) / BLOCK_SIZE_BYTES))
+    block_count = ceil(float(len(data)) / BLOCK_SIZE_BYTES)
 
     encrypted_data = []
     previous_cipher_block = iv
@@ -109,7 +109,7 @@ def aes_ecb_encrypt(data, key):
     @returns {int[]}           encrypted data
     """
     expanded_key = key_expansion(key)
-    block_count = int(ceil(float(len(data)) / BLOCK_SIZE_BYTES))
+    block_count = ceil(float(len(data)) / BLOCK_SIZE_BYTES)
 
     encrypted_data = []
     for i in range(block_count):
